feat:add user and role

+++ routes/web.php
@@ -2,6 +2,7 @@
 
 use App\Http\Controllers\MainController;
 use App\Http\Controllers\AdminDashboardController;
+use App\Http\Controllers\AuthController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -15,8 +16,8 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-// Route::get('/', function () {
-//     return view('home');
+// Route::get('/home', function () {
+//     return redirect()->to('/login');
 // });
 Route::get('/', [MainController::class, 'index']);
 Route::get('/congress-information', [MainController::class, 'congressInfo']);
@@ -28,23 +29,35 @@ Route::get('/submission', [MainController::class, 'submission']);
 //     return view('dashboard.admin.home');
 // });
 Route::prefix('/dashboard/admin')->group(function () {
-    Route::get('/', [AdminDashboardController::class, 'index']);
-    Route::get('/user-register', [AdminDashboardController::class, 'userRegister']);
-    Route::get('/user-role', [AdminDashboardController::class, 'role']);
-    Route::get('/important-dates', [AdminDashboardController::class, 'importantDates']);
-    Route::post('/storeImportantDates', [AdminDashboardController::class, 'storeImportantDates']);
-    Route::delete('/delete_importantDate/{id}', [AdminDashboardController::class, 'delete_importantDate']);
-    Route::get('/important-dates/{id}/restore', [AdminDashboardController::class, 'restore_importantDate']);
-    Route::get('/edit-important-dates/{id}', [AdminDashboardController::class, 'edit_importantDate']);
-    Route::put('/update_importantDate/{id}', [AdminDashboardController::class, 'update_importantDate']);
-    Route::get('/faculties', [AdminDashboardController::class, 'faculties']);
-    Route::post('/storeFaculties', [AdminDashboardController::class, 'storeFaculties']);
-    Route::delete('/delete_faculty/{id}', [AdminDashboardController::class, 'delete_faculty']);
-    Route::get('/faculties/{id}/restore', [AdminDashboardController::class, 'restore_faculty']);
-    Route::get('/edit_faculties/{id}', [AdminDashboardController::class, 'edit_faculties']);
-    Route::put('/update_faculties/{id}', [AdminDashboardController::class, 'update_faculties']);
-    Route::get('/schedule', [AdminDashboardController::class, 'schedules']);
-    Route::post('/storeSchedule', [AdminDashboardController::class, 'storeSchedule']);
-    Route::delete('delete_schedule/{id}', [AdminDashboardController::class, 'delete_schedule']);
-    Route::get('/schedule/{id}/restore', [AdminDashboardController::class, 'restore_schedule']);
+    Route::get('/', [AdminDashboardController::class, 'index'])->middleware('auth', 'verified');
+    Route::get('/important-dates', [AdminDashboardController::class, 'importantDates'])->middleware('auth', 'verified');
+    Route::post('/storeImportantDates', [AdminDashboardController::class, 'storeImportantDates'])->middleware('auth', 'verified');
+    Route::delete('/delete_importantDate/{id}', [AdminDashboardController::class, 'delete_importantDate'])->middleware('auth', 'verified');
+    Route::get('/important-dates/{id}/restore', [AdminDashboardController::class, 'restore_importantDate'])->middleware('auth', 'verified');
+    Route::get('/edit-important-dates/{id}', [AdminDashboardController::class, 'edit_importantDate'])->middleware('auth', 'verified');
+    Route::put('/update_importantDate/{id}', [AdminDashboardController::class, 'update_importantDate'])->middleware('auth', 'verified');
+    Route::get('/faculties', [AdminDashboardController::class, 'faculties'])->middleware('auth', 'verified');
+    Route::post('/storeFaculties', [AdminDashboardController::class, 'storeFaculties'])->middleware('auth', 'verified');
+    Route::delete('/delete_faculty/{id}', [AdminDashboardController::class, 'delete_faculty'])->middleware('auth', 'verified');
+    Route::get('/faculties/{id}/restore', [AdminDashboardController::class, 'restore_faculty'])->middleware('auth', 'verified');
+    Route::get('/edit_faculties/{id}', [AdminDashboardController::class, 'edit_faculties'])->middleware('auth', 'verified');
+    Route::put('/update_faculties/{id}', [AdminDashboardController::class, 'update_faculties'])->middleware('auth', 'verified');
+    Route::get('/schedule', [AdminDashboardController::class, 'schedules'])->middleware('auth', 'verified');
+    Route::post('/storeSchedule', [AdminDashboardController::class, 'storeSchedule'])->middleware('auth', 'verified');
+    Route::delete('delete_schedule/{id}', [AdminDashboardController::class, 'delete_schedule'])->middleware('auth', 'verified');
+    Route::get('/schedule/{id}/restore', [AdminDashboardController::class, 'restore_schedule'])->middleware('auth', 'verified');
+    Route::get('/user-register', [AuthController::class, 'userRegister'])->middleware('auth', 'verified');
+    Route::post('/storeUser', [AuthController::class, 'storeUser'])->name('storeUser')->middleware('auth', 'verified');
+    Route::get('/user-role', [AuthController::class, 'role'])->middleware('auth', 'verified');
+    Route::post('/storeRole', [AuthController::class, 'storeRole'])->name('storeRole')->middleware('auth', 'verified');
+    Route::delete('/delete_role/{id}', [AuthController::class, 'delete_role'])->middleware('auth', 'verified');
+    Route::get('/user-role/{id}/restore', [AuthController::class, 'restore_role'])->middleware('auth', 'verified');
 });
+
+
+Route::group(['middleware' => 'guest'], function () {
+    Route::get('/login', [AuthController::class, 'login'])->name('login');
+    Route::post('/login', [AuthController::class, 'storeLogin'])->name('login');
+});
+
+Route::post('/logout', [AuthController::class, 'logout'])->name('logout');
